+++ chocolatey-visualstudio.extension/extensions/Get-VSProductReference.ps1
@@ -3,25 +3,11 @@
     [CmdletBinding()]
     Param
     (
-        [Parameter(Mandatory = $true)] [string] $VisualStudioYear,
+        [Parameter(Mandatory = $true)] [PSObject] $ChannelReference,
         [Parameter(Mandatory = $true)] [string] $Product
     )
 
-    switch ($VisualStudioYear)
-    {
-        '2017' { $channelId = 'VisualStudio.15.Release' }
-        default { throw "Unsupported VisualStudioYear: $VisualStudioYear"}
-    }
-
     $productId = "Microsoft.VisualStudio.Product." + $Product
-
-    $props = @{
-        ChannelId = $channelId
-        ChannelUri = $null
-        InstallChannelUri = $null
-        ProductId = $productId
-    }
-
-    $obj = New-Object -TypeName PSObject -Property $props
+    $obj = New-VSProductReference -ChannelId $ChannelReference.ChannelId -ProductId $productId
     return $obj
 }