+++ dotnetfx/tools/ChocolateyInstall.ps1
@@ -1,11 +1,11 @@
 ﻿. (Join-Path -Path (Split-Path -Parent -Path $MyInvocation.MyCommand.Definition) -ChildPath 'helpers.ps1')
 
 $packageName = 'dotnetfx'
-$release = 527934
+$release = 528033
 $version = '4.8'
-$productNameWithVersion = "Microsoft .NET Framework $version early access build 3646"
-$url = 'https://download.microsoft.com/download/6/5/7/6577634A-8D5D-4558-BA22-A81CC6D5BB06/NDP48-Preview-x86-x64-AllOS-ENU.exe'
-$checksum = '5D6F819B273D097586286284FCB654B3400A41C446062CF3E426C1FB92F9119A'
+$productNameWithVersion = "Microsoft .NET Framework $version early access build 3745"
+$url = 'https://download.visualstudio.microsoft.com/download/pr/9854b5f2-2341-4136-ad7d-1d881ab8d603/a59983c9bf08c7ae7200f356c9a604c1/NDP48-Preview-x86-x64-AllOS-ENU.exe'
+$checksum = 'D1C399178BF9733BC56C223F212512879E6ECDD6C647F25A1589C86C05591281'
 $checksumType = 'sha256'
 
 if (Test-Installed -Release $release) {