+++ visualstudio2017-workload-universalbuildtools/tools/ChocolateyUninstall.ps1
@@ -1,0 +1,5 @@
+﻿Remove-VisualStudioWorkload `
+    -PackageName 'visualstudio2017-workload-universalbuildtools' `
+    -Workload 'UniversalBuildTools' `
+    -VisualStudioYear '2017' `
+    -ApplicableProducts @('BuildTools')