+++ visualstudio2017buildtools/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017buildtools' `
     -ApplicationName 'Microsoft Visual Studio Build Tools 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/d80c9e2f-b6f4-47cc-bc8b-1bb40ec4c92d/7189a68796aed20aabb13985c49d530b/vs_buildtools.exe' `
-    -Checksum '37EFB023D59CBC4B6E828EFE3EA2A5D1B36344475659ED2F4DC8DC09FF2811C4' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/f8db94a7-42dc-4413-b987-b21dcb0b1661/5e24cf2c187f46bde25c5515250fa9a0/vs_buildtools.exe' `
+    -Checksum 'F5162FDEF68BC22DBCB52A6558725AED951B0ACE15885E526817831EB6F5BA72' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'BuildTools' `