+++ visualstudio2017community/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017community' `
     -ApplicationName 'Microsoft Visual Studio Community 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/cde51e45-6536-4293-88c4-9dc341a3eb84/fe7096403605291306155738116c3153/vs_community.exe' `
-    -Checksum '383CF3B83A2DCA935967753652C67A287B6B3BA391EF6F5979540C18A1475486' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/41217ce6-f73c-48b9-b679-e5193984336b/500a2965365fa0283c3c31e4837487d9/vs_community.exe' `
+    -Checksum '74B519B94EA7655E4ED38083F0E08B0FBF3AE0CAC5C551719B7086D20D6E999A' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'Community' `