+++ visualstudio2017enterprise/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017enterprise' `
     -ApplicationName 'Microsoft Visual Studio Enterprise 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/a7dca6fc-eb4f-4be8-9fa3-10941f5de7ce/5a45e2fb181171d68f139c06818d01dd/vs_enterprise.exe' `
-    -Checksum 'EFB11F71935BE04F7DE042C6725938803E36BDF073C27A4682A1FD32129AA551' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/324e8588-c90a-4229-947f-d60efec43018/0baa0fb3754413e8048d3625ddf5d585/vs_enterprise.exe' `
+    -Checksum 'B99B6810AB453EC8D99C3F833532F9E07A28A575672AE199E5E1A27B52CA5E7E' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'Enterprise' `