+++ visualstudio2017feedbackclient/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017feedbackclient' `
     -ApplicationName 'Microsoft Visual Studio Feedback Client 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/9893f8ca-f173-41eb-ae29-4db6c4aa0bf0/148f1024345f35371ab41ee781509903/vs_feedbackclient.exe' `
-    -Checksum 'DD1445053582E0F4254553B7E642757BE067392FB8DFDC8E794B9F46CA60E72A' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/e1f60dd7-7df1-4ebe-bea3-b303b575dd3b/4a133bc2f4d11c7f3ee1e32b79b4b09e/vs_feedbackclient.exe' `
+    -Checksum 'ED6A7318727D42CB88D98102661C24F2E922195C0B97B42BC4CCF281F33F3E13' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'FeedbackClient' `