+++ visualstudio2017professional/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017professional' `
     -ApplicationName 'Microsoft Visual Studio Professional 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/a401cb72-762b-409d-b7cc-5ac9a6bd67db/be7a0d6c6c80187a8c7c4a43a81473f8/vs_professional.exe' `
-    -Checksum 'E8C5B061759DE7E1F9BD1712B4518E7D81876B99A83D6A25E6F048F0D14B7341' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/366df54c-35a5-4d91-9614-fc0044349ed7/dc65bed6acbf3e2bd2d3c388c57b9987/vs_professional.exe' `
+    -Checksum '0ED75BA80D5A578B21C2BDC13C6C04514632AC5823E9F7642F385C280445E4B1' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'Professional' `