+++ visualstudio2017testagent/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017testagent' `
     -ApplicationName 'Microsoft Visual Studio Test Agent 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/2436bbfc-0dc4-4328-9b94-c8a11f0a7341/208858a0e660b67f065e47819d36750b/vs_testagent.exe' `
-    -Checksum '642A4BC5843EE0C3622E9514D9AF734F473D0DCD3923F706EBD523953F95FA8D' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/4a52bd7f-6921-4cd5-a054-8f45a90bc933/0e4b6ce290ab9ad0c174e8fbb99b0929/vs_testagent.exe' `
+    -Checksum 'EFD06837C13A60FC2B8F4FE76043C84AF021D9417F7EA3EADAB2FB4F0A55C0E6' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'TestAgent' `