+++ visualstudio2017testcontroller/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017testcontroller' `
     -ApplicationName 'Microsoft Visual Studio Test Controller 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/600a87d9-0ae6-4ac0-b783-e3221eb99421/0aac6e00889f436fcece4b301137fb93/vs_testcontroller.exe' `
-    -Checksum 'EF910F86BF5B692934914319646A93555E219010D2881C60A4F93712D7474049' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/0ca6d732-21bd-43c4-95fc-e8a5a389d9e4/5b672f904d23846eede6b34f5cea60f2/vs_testcontroller.exe' `
+    -Checksum 'EBDC405AFD7864CF7D16CDE4EA49C05A08324F7FE555F7A5225F429B1621C0B0' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'TestController' `