+++ visualstudio2017testprofessional/tools/ChocolateyInstall.ps1
@@ -1,8 +1,8 @@
 ﻿Install-VisualStudio `
     -PackageName 'visualstudio2017testprofessional' `
     -ApplicationName 'Microsoft Visual Studio Test Professional 2017' `
-    -Url 'https://download.visualstudio.microsoft.com/download/pr/5a31abc8-2448-4b9f-a961-82800a083d47/35806c3192fbdda58c362540211bdea0/vs_testprofessional.exe' `
-    -Checksum 'BB59E7AFC818054710FEDFA3F07482A6E6797310EB2AD1CE1C46DC48B28A31F8' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/e898f48a-4628-40ca-a9bd-c3ffd13830d9/e4e1a436ea33df791299ae2d1c25a58f/vs_testprofessional.exe' `
+    -Checksum '416146CF8179869CB31B4DA7A6265CBD58F41B83C9F3CA49B541647FF35C3B4A' `
     -ChecksumType 'SHA256' `
     -InstallerTechnology 'WillowVS2017OrLater' `
     -Product 'TestProfessional' `