+++ visualstudio2019buildtools-preview/tools/ChocolateyInstall.ps1
@@ -1,0 +1,10 @@
+﻿Install-VisualStudio `
+    -PackageName 'visualstudio2019buildtools-preview' `
+    -ApplicationName 'Microsoft Visual Studio Build Tools 2019 Preview' `
+    -Url 'https://download.visualstudio.microsoft.com/download/pr/f857c723-17a5-4dcb-9180-17fe9b787511/d75909023037747f8f3b8c66ce4d05dd/vs_buildtools.exe' <# https://aka.ms/vs/16/pre/vs_buildtools.exe #> `
+    -Checksum '90F74908E882779F65EE5A54083D26B5BA4CBF3B029E6632F6A922D2E3944422' `
+    -ChecksumType 'SHA256' `
+    -InstallerTechnology 'WillowVS2017OrLater' `
+    -Product 'BuildTools' `
+    -VisualStudioYear '2019' `
+    -Preview $true